upd golang version

+++ pkg/utils/log.go
@@ -59,14 +59,14 @@ func NewLogrusLogger(level log.Level, prefix string, fields log.Fields) log.Logg
 		Logger: logger,
 		level:  level,
 	}
-	logger.SetLevel(level)
+	logger.SetLevel(uint32(level))
 	return logger.WithPrefix(prefix)
 }
 
 func SetLogLevel(prefix string, level log.Level) error {
 	if logger, found := loggers[prefix]; found {
 		logger.level = level
-		logger.Logger.SetLevel(level)
+		logger.Logger.SetLevel(uint32(level))
 		return nil
 	}
 	return fmt.Errorf("logger [%v] not found", prefix)
